requirements.txt
fastapi
uvicorn[standard]
requests
# numpy
# torch
# scipy
opencv-python
ultralytics[export]
timm
pydantic-settings
python-multipart
pyrealsense2
python-magic